[MAP]: Show map layers in web page
The map layers (currently only the base image and generated database tile layer) are layered on top of each other in a HTML canvas web page. This could have been easily done with a Python script, but I chose HTML canvas because it's easily expandable to complex interactivity and works on most systems without installing anything.

The web page and images have been placed in the folder "map".

The map tool is now still in a bodged first concept phase. Many parameters such as file names, image resolutions and map layers are still embedded in the code. If a more advanced map tool is required, this will probably have to be done in a more elegantly structured way, but for now it will do.