[COUPLE]: Use wget --reject as file filter
Instead of saving and keeping a list of coordinates that had any HDF5 files in the local database associated with them, couple_collect.py now scans the ./tiles folder for filenames to reject. These tiff filenames correspond with the original HDF5 filename. This allows for additional HDF5 downloads for coordinates that already have some files associated with them. All the filenames get concatenated into one big string to feed as an argument to wget --reject.

This also means couple.py no longer needs to save a list of processed coordinates.