[COUPLE]: Bitshift for nicer viewing
All of the image data values seem to cover only a fraction (~2^10) of the uint16 datatype maximum (2^16-1). This causes the images to appear black when viewed by a simple image viewer. Though this matters not for scientific purposes, it is slightly nicer to be able to see the images you're working with. In section 5.1 of the PROBA-V manual, it is seems to be mentioned that only 12 bits are actually used. Therefore, a bitshift of 4 bits should do no harm to the data, while making the image previews much nicer for humans.