[BUILD]: Make and include Score Mask
The Score Mask (SM) indicates which pixels can (in theory) be resolved with SR and checked against the HR image; i.e. there should be at least one LR where the pixel is clear, and it should be clear in the HR. If not, the pixel will be marked black in Score Mask.

Pixels marked black in the Score Mask will be ignored during the scoring process. Note that due to water vapour, most rivers and lakes will be marked as black in the Score Mask.