[BUILD]: Add validate and submission-test folder
The validate folder will contain all files required for validation of a submission. It currently holds the HR files, normalization factors (in text files) and the Score Masks.

The submission-test folder will contain the HR files with added noise, in with the correct file names. This can be used as a test for the submission script on the Kelvins website.